ENH:烧录大小限定在size-12是8的倍数，优化跳转函数

+++ Utilities/inc/bsp_rs485.h
@@ -44,7 +44,7 @@
 
 /* 全局变量 */
 extern uint32_t rs485_rx_count;
-extern uint32_t rs485_rxbuffer[RS485_BUFFER_SIZE];
+extern uint8_t rs485_rxbuffer[RS485_BUFFER_SIZE];
 extern __IO FlagStatus rs485_idle_flag;
 extern __IO FlagStatus rs485_tx_flag;
